started working with edgepaths
