+++ superficial/src/main/scala/superficial/TwoComplex.scala
@@ -664,88 +664,23 @@ trait TwoComplex { twoComplex =>
   /**
     * Vector of edges to the left of an edge, modified version of orbit
     */
-    def vectorOrbit (e : Edge, opt: (Edge => Option[Edge]), accum : Vector[Edge]) : Vector[Edge] = {
-        val nextEdge = opt(e)
-        if ((nextEdge != None) && (! accum.contains(nextEdge))) vectorOrbit(nextEdge.get, opt, accum :+ nextEdge.get)
-        else accum.map(_.flip)
-      }
+  def vectorOrbit (e : Edge, opt: (Edge => Option[Edge]), accum : Vector[Edge]) : Vector[Edge] = {
+    val nextEdge = opt(e)
+    if ((nextEdge != None) && (! accum.contains(nextEdge))) vectorOrbit(nextEdge.get, opt, accum :+ nextEdge.get)
+    else accum.map(_.flip)
+  }
   
-    /**
-     * Vector of edges to the left of an edge
-     */
+  /**
+  * Vector of edges to the left of an edge
+  */
 
-     def vectorEdgesToTheLeftOf(e: Edge) = vectorOrbit(e, rotateLeftOpt(_), Vector[Edge]())
+  def vectorEdgesToTheLeftOf(e: Edge) = vectorOrbit(e, rotateLeftOpt(_), Vector[Edge]())
 
-     /**
-       * Vector of edges to the right of an edge
-       */
+  /**
+  * Vector of edges to the right of an edge
+  */
 
-    def vectorEdgesToTheRightOf(e: Edge) = vectorOrbit(e, rotateRightOpt(_), Vector[Edge]())
-
-// --------------------------------------------------------------------------------------------------------------------
-  
-  // trial and error area -----------------------------------------------------------  
-      
-// ----------------------------------------------------------------------------------------------------------
-  def quadrangulate : TwoComplex = {
- 
-    require(twoComplex.isClosedSurface, "Algorithm only works for closed surfaces")       
-
-    val faceList = twoComplex.faces.toList
-    val facesWithIndexes :  List[(Polygon, Int)] = faceList.flatMap(f => ((0 to (f.boundary.length - 1)).map(ind => (f, ind))))
-
-    def createBarycenter (face : Polygon) : Vertex = {
-      object bFace extends Vertex
-      bFace
-    } 
-
-    val barycentersList = faceList.map(createBarycenter(_))
-    val barycenters = faceList.zip(barycentersList).toMap
-
-    def createEdgePairs (face : Polygon, index : Int) : EdgePair = {
-      new EdgePair(barycenters(face), face.boundary(index).terminal)
-    }
-
-    val newEdgeMap1 : Map[(Polygon, Int),EdgePair] = 
-      facesWithIndexes.map{
-        case (poly, j) => (poly, j) -> createEdgePairs(poly, j)
-      }.toMap
-
-    def mod (m : Int, n : Int) = ((m % n) + n) % n 
-  
-    val vertexList = twoComplex.vertices.toList
-    val edgeList = twoComplex.edges.toList
-
-    def faceOfEdge (edge : Edge) : Polygon = {
-       val face = faces.find(_.boundary.contains(edge))
-       assert(face != None, "For a closed surface each edge should be in at least one face")
-       face.get
-    }
-    
-    def createFace (edge : Edge) : Polygon = {
-      val face = faceOfEdge(edge)
-      val flipFace = faceOfEdge(edge.flip)
-      val indexOfEdge = face.boundary.indexOf(edge)
-      val indexOfFlip = flipFace.boundary.indexOf(edge.flip)
-      val periOfFace = face.boundary.length
-      val periOfFlip = flipFace.boundary.length
-
-
-
-      Polygon.apply(Vector(
-        newEdgeMap1(face, indexOfEdge).Positive,
-        newEdgeMap1(flipFace, mod(indexOfFlip - 1, periOfFlip)).Negative,
-        newEdgeMap1(flipFace, indexOfFlip).Positive,
-        newEdgeMap1(face, mod(indexOfEdge - 1, periOfFace)).Negative
-      ))
-    }  
-
-    val newFaces = halfEdges.map(createFace)
-    object quad extends PureTwoComplex {
-      val faces = newFaces
-    }
-    quad
-  }  
+  def vectorEdgesToTheRightOf(e: Edge) = vectorOrbit(e, rotateRightOpt(_), Vector[Edge]())    
 }
